+++ src/main/kotlin/com/deepromeet/atcha/transit/domain/ServiceRegionValidator.kt
@@ -1,0 +1,15 @@
+package com.deepromeet.atcha.transit.domain
+
+import com.deepromeet.atcha.location.domain.Coordinate
+import org.springframework.stereotype.Component
+
+@Component
+class ServiceRegionValidator(
+    private val regionIdentifier: RegionIdentifier
+) {
+    fun validate(vararg coordinates: Coordinate) {
+        coordinates.forEach { coordinate ->
+            regionIdentifier.identify(coordinate)
+        }
+    }
+}
